requirements.txt
Django==5.0.6
numpy==1.26.3
sentencepiece==0.2.0

djangorestframework~=3.15.1
torch~=2.3.1+cu121
Markdown~=3.6
beautifulsoup4~=4.12.3